Split test helpers from tests of helper functionality.
