+++ test/definitions/gcn.jl
@@ -1,13 +1,13 @@
 using GPUCompiler
 
 if !@isdefined(TestRuntime)
-    include("../util.jl")
+    include("../testhelpers.jl")
 end
 
 
 # create a GCN-based test compiler, and generate reflection methods for it
 
-function gcn_job(@nospecialize(func), @nospecialize(types); 
+function gcn_job(@nospecialize(func), @nospecialize(types);
                  kernel::Bool=false, always_inline=false, kwargs...)
     source = FunctionSpec(typeof(func), Base.to_tuple_type(types), kernel)
     target = GCNCompilerTarget(dev_isa="gfx900")
